refactor: Методы клиента отладки выделены в отдкльный класс
