+++ src/main/java/com/clouds42/Commands/CoverageCommand.java
@@ -21,24 +21,11 @@
  */
 package com.clouds42.Commands;
 
-import com._1c.g5.v8.dt.debug.core.runtime.client.RuntimeDebugClientException;
-import com._1c.g5.v8.dt.debug.model.base.data.AttachDebugUIResult;
-import com._1c.g5.v8.dt.debug.model.base.data.BSLModuleIdInternal;
-import com._1c.g5.v8.dt.debug.model.base.data.DebugTargetId;
-import com._1c.g5.v8.dt.debug.model.dbgui.commands.DBGUIExtCmdInfoBase;
-import com._1c.g5.v8.dt.debug.model.dbgui.commands.DBGUIExtCmds;
-import com._1c.g5.v8.dt.debug.model.dbgui.commands.impl.DBGUIExtCmdInfoMeasureImpl;
-import com._1c.g5.v8.dt.debug.model.dbgui.commands.impl.DBGUIExtCmdInfoStartedImpl;
-import com._1c.g5.v8.dt.debug.model.measure.PerformanceInfoLine;
-import com._1c.g5.v8.dt.debug.model.measure.PerformanceInfoMain;
-import com._1c.g5.v8.dt.debug.model.measure.PerformanceInfoModule;
-import com._1c.g5.v8.dt.internal.debug.core.runtime.client.RuntimeDebugModelXmlSerializer;
 import com.clouds42.CommandLineOptions.*;
-import com.clouds42.DebugClient;
-import com.clouds42.MyRuntimeDebugModelXmlSerializer;
 import com.clouds42.PipeMessages;
 import com.clouds42.Utils;
-import org.eclipse.emf.common.util.EList;
+import com.github._1c_syntax.coverage41C.DebugClientException;
+import com.github._1c_syntax.coverage41C.EDT.DebugClientEDT;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
@@ -52,12 +39,9 @@ import java.lang.invoke.MethodHandles;
 import java.lang.module.ModuleDescriptor.Version;
 import java.math.BigDecimal;
 import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
 import java.util.concurrent.Callable;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.stream.Stream;
 
 @Command(name = PipeMessages.START_COMMAND, mixinStandardHelpOptions = true,
         description = "Start measure and save coverage data to file",
@@ -87,7 +71,7 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
     @Option(names = {"--opid"}, description = "Owner process PID", defaultValue = "-1")
     Integer opid;
 
-    private DebugClient client;
+    private DebugClientEDT client;
 
     private final Map<URI, Map<BigDecimal, Integer>> coverageData = new HashMap<>() {
         @Override
@@ -103,23 +87,7 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
 
 
     private final AtomicBoolean stopExecution = new AtomicBoolean(false);
-    private boolean rawMode = false;
     private boolean systemStarted = false;
-
-    private void connectAllTargets(List<DebugTargetId> debugTargets) {
-        logger.info("Current debug targets size: {}", debugTargets.size());
-        debugTargets.forEach(debugTarget -> {
-            String id = debugTarget.getId();
-            String seanceId = debugTarget.getSeanceId();
-            String targetType = debugTarget.getTargetType().getName();
-            logger.info("Id: {} , seance id: {} , target type: {}", id, seanceId, targetType);
-            try {
-                client.attachRuntimeDebugTargets(Collections.singletonList(UUID.fromString(debugTarget.getId())));
-            } catch (RuntimeDebugClientException e) {
-                logger.error(e.getLocalizedMessage());
-            }
-        });
-    }
 
     @Override
     public Integer call() throws Exception {
@@ -127,19 +95,26 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
         int result = CommandLine.ExitCode.OK;
         getServerSocket();
 
+        client = new DebugClientEDT();
+        client.setOptions(
+                loggingOptions.isVerbose(),
+                metadataOptions.isRawMode()
+        );
 
-        RuntimeDebugModelXmlSerializer serializer = new MyRuntimeDebugModelXmlSerializer();
-        client = new DebugClient(serializer);
+        client.setResolverOptions(
+                filterOptions.getExtensionName(),
+                filterOptions.getExternalDataProcessorUrl(),
+                coverageData
+        );
+
 
         UUID measureUuid = UUID.randomUUID();
-
-        rawMode = metadataOptions.isRawMode();
 
         Map<String, URI> uriListByKey = Utils.readMetadata(metadataOptions, coverageData);
 
         try {
             startSystem(measureUuid);
-        } catch (RuntimeDebugClientException e) {
+        } catch (DebugClientException e) {
             logger.info("Connecting to dbgs failed");
             logger.error(e.getLocalizedMessage());
             result = CommandLine.ExitCode.SOFTWARE;
@@ -152,7 +127,7 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
 
         try {
             mainLoop(uriListByKey, externalDataProcessorsUriSet);
-        } catch (RuntimeDebugClientException e) {
+        } catch (DebugClientException e) {
             logger.error("Can't send ping to debug server. Coverage analyzing finished");
             logger.error(e.getLocalizedMessage());
             e.printStackTrace();
@@ -171,13 +146,7 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
 
         gracefulShutdown(null);
 
-        logger.info("Disconnecting from dbgs...");
-        try {
-            client.disconnect();
-            client.dispose();
-        } catch (RuntimeDebugClientException e) {
-            logger.error(e.getLocalizedMessage());
-        }
+        client.disconnect();
         closeSocket();
 
         logger.info("Main thread finished");
@@ -195,136 +164,27 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
         }));
     }
 
-    private void targetStarted(DBGUIExtCmdInfoStartedImpl command) {
-        DebugTargetId targetId = command.getTargetID();
-        try {
-            client.attachRuntimeDebugTargets(Collections.singletonList(UUID.fromString(targetId.getId())));
-        } catch (RuntimeDebugClientException e) {
-            logger.info("Command: {} error!", command.getCmdID().getName());
-            logger.error(e.getLocalizedMessage());
-        }
-    }
-
-    private void mainLoop(Map<String, URI> uriListByKey, Set<String> externalDataProcessorsUriSet) throws RuntimeDebugClientException {
+    private void mainLoop(Map<String, URI> uriListByKey, Set<String> externalDataProcessorsUriSet) throws DebugClientException {
         while (!stopExecution.get()) {
-            List<? extends DBGUIExtCmdInfoBase> commandsList = client.ping();
-            logger.info("Ping result commands size: {}", commandsList.size());
-            commandsList.forEach(command -> {
-                logger.info("Command: {}", command.getCmdID().getName());
-                if (command.getCmdID() == DBGUIExtCmds.MEASURE_RESULT_PROCESSING) {
-                    measureResultProcessing(uriListByKey, externalDataProcessorsUriSet, (DBGUIExtCmdInfoMeasureImpl) command);
-                } else if (command.getCmdID() == DBGUIExtCmds.TARGET_STARTED) {
-                    targetStarted((DBGUIExtCmdInfoStartedImpl) command);
-                }
-            });
+            client.ping(uriListByKey, externalDataProcessorsUriSet);
         }
     }
 
-    private void measureResultProcessing(Map<String, URI> uriListByKey, Set<String> externalDataProcessorsUriSet, DBGUIExtCmdInfoMeasureImpl command) {
-        logger.info("Found MEASURE_RESULT_PROCESSING command");
-        PerformanceInfoMain measure = command.getMeasure();
-        EList<PerformanceInfoModule> moduleInfoList = measure.getModuleData();
-        moduleInfoList.forEach(moduleInfo -> {
-            BSLModuleIdInternal moduleId = moduleInfo.getModuleID();
-            String moduleUrl = moduleId.getURL();
-            if (loggingOptions.isVerbose() && !moduleUrl.isEmpty() && !externalDataProcessorsUriSet.contains(moduleUrl)) {
-                logger.info("Found external data processor: {}", moduleUrl);
-                externalDataProcessorsUriSet.add(moduleUrl);
-            }
-            String moduleExtensionName = moduleId.getExtensionName();
-            if (rawMode
-                    || (filterOptions.getExtensionName().equals(moduleExtensionName)
-                    && filterOptions.getExternalDataProcessorUrl().equals(moduleUrl))) {
-                String objectId = moduleId.getObjectID();
-                String propertyId = moduleId.getPropertyID();
-                String key = Utils.getUriKey(objectId, propertyId);
-
-                URI uri;
-                if (!rawMode) {
-                    uri = uriListByKey.get(key);
-                } else {
-                    uri = URI.create("file:///" + key);
-                }
-                if (uri == null) {
-                    logger.info("Couldn't find object id {}, property id {} in sources!", objectId, propertyId);
-                } else {
-                    EList<PerformanceInfoLine> lineInfoList = moduleInfo.getLineInfo();
-                    lineInfoList.forEach(lineInfo -> {
-                        BigDecimal lineNo = lineInfo.getLineNo();
-                        Map<BigDecimal, Integer> coverMap = coverageData.get(uri);
-                        if (!coverMap.isEmpty() || rawMode) {
-                            if (!rawMode && !coverMap.containsKey(lineNo)) {
-                                if (loggingOptions.isVerbose()) {
-                                    logger.info("Can't find line to cover {} in module {}", lineNo, uri);
-                                    try {
-                                        Stream<String> all_lines = Files.lines(Paths.get(uri));
-                                        Optional<String> first = all_lines.skip(lineNo.longValue() - 1).findFirst();
-                                        if (first.isPresent()) {
-                                            String specific_line_n = first.get();
-                                            logger.info(">>> {}", specific_line_n);
-                                        }
-                                    } catch (Exception e) {
-                                        logger.error(e.getLocalizedMessage());
-                                    }
-                                }
-                            } else {
-                                int currentValue = coverMap.getOrDefault(lineNo, 0);
-                                if (currentValue < 0) {
-                                    currentValue = 0;
-                                }
-                                coverMap.put(lineNo,
-                                        currentValue
-                                                + lineInfo.getFrequency().intValue());
-                            }
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    private void startSystem(UUID measureUuid) throws RuntimeDebugClientException {
-        UUID debugServerUuid = UUID.randomUUID();
+    private void startSystem(UUID measureUuid) throws DebugClientException {
         client.configure(
                 connectionOptions.getDebugServerUrl(),
-                debugServerUuid,
                 connectionOptions.getInfobaseAlias());
-        logger.info("Connecting to debugger...");
-        AttachDebugUIResult connectionResult = client.connect(debuggerOptions.getPassword());
-        if (connectionResult != AttachDebugUIResult.REGISTERED) {
-            if (connectionResult == AttachDebugUIResult.IB_IN_DEBUG) {
-                throw new RuntimeDebugClientException("Can't connect to debug server. IB is in debug. Close configurator or EDT first");
-            } else if (connectionResult == AttachDebugUIResult.CREDENTIALS_REQUIRED) {
-                throw new RuntimeDebugClientException("Can't connect to debug server. Use -p option to set correct password");
-            } else {
-                throw new RuntimeDebugClientException("Can't connect to debug server. Connection result: " + connectionResult);
-            }
-        }
+
+        client.connect(debuggerOptions.getPassword());
+
         Version apiver = Version.parse(client.getApiVersion());
-        logger.info("Setup settings...");
-        client.initSettings(false);
-        client.setAutoconnectDebugTargets(
+
+        client.setupSettings(
                 debuggerOptions.getDebugAreaNames(),
                 debuggerOptions.getFilteredAutoconnectTargets(apiver));
-        logger.info("Setup targets...");
-        List<DebugTargetId> debugTargets;
-        if (debuggerOptions.getDebugAreaNames().isEmpty()) {
-            debugTargets = client.getRuntimeDebugTargets(null);
-        } else {
-            debugTargets = new LinkedList<>();
-            debuggerOptions.getDebugAreaNames().forEach(areaName -> {
-                try {
-                    debugTargets.addAll(client.getRuntimeDebugTargets(areaName));
-                } catch (RuntimeDebugClientException ex) {
-                    logger.error(ex.getLocalizedMessage());
-                }
-            });
-        }
-        connectAllTargets(debugTargets);
 
-        logger.info("Enabling profiling...");
-        client.toggleProfiling(null);
-        client.toggleProfiling(measureUuid);
+        client.connectTargets(debuggerOptions);
+        client.enableProfiling(measureUuid);
 
         systemStarted = true;
 
@@ -335,12 +195,7 @@ public class CoverageCommand extends CoverServer implements Callable<Integer> {
             return;
         }
 
-        logger.info("Disabling profiling...");
-        try {
-            client.toggleProfiling(null);
-        } catch (RuntimeDebugClientException e) {
-            logger.error(e.getLocalizedMessage());
-        }
+        client.disableProfiling();
 
         Utils.dumpCoverageFile(coverageData, metadataOptions, outputOptions);
         if (serverPipeOut != null) {
