+++ src/main/java/com/github/_1c_syntax/coverage41C/DebugClientException.java
@@ -1,0 +1,12 @@
+package com.github._1c_syntax.coverage41C;
+
+public class DebugClientException extends Exception {
+
+    public DebugClientException(String message) {
+        super(message);
+    }
+
+    public DebugClientException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
